Initial cleanup

+++ firmware/src/motor/calibration.c
@@ -245,7 +245,7 @@ bool calibrate_offset_and_rectification(void)
     wait_a_while();
     int16_t offset_raw = ma7xx_get_angle_raw();
     // Perform measuerments, store only mean of F + B error
-    for (uint32_t i = 0; i < n; i++)
+    for (int16_t i = 0; i < n; i++)
     {
         for (uint8_t j = 0; j < nconv; j++)
         {
@@ -257,7 +257,7 @@ bool calibrate_offset_and_rectification(void)
         error_ticks[i] = (int16_t)(e_pos_ref * e_pos_to_ticks - pos_meas);
     }
     offset_raw = (offset_raw + ma7xx_get_angle_raw()) / 2;
-    for (uint32_t i = 0; i < n; i++)
+    for (int16_t i = 0; i < n; i++)
     {
         for (uint8_t j = 0; j < nconv; j++)
         {
